added custom exception class for review entity.

+++ library-project/src/main/java/com/capgemini/library_project/services/BorrowRecordServicesImpl.java
@@ -9,8 +9,11 @@ import com.capgemini.library_project.entities.Book;
 import com.capgemini.library_project.entities.BorrowRecord;
 import com.capgemini.library_project.entities.User;
 import com.capgemini.library_project.exceptions.AlreadyReturnedException;
+import com.capgemini.library_project.exceptions.BookNotFoundException;
+import com.capgemini.library_project.exceptions.BorrowRecordNotFoundException;
 import com.capgemini.library_project.exceptions.InvalidBorrowDateException;
 import com.capgemini.library_project.exceptions.InvalidStatusException;
+import com.capgemini.library_project.exceptions.UserNotFoundException;
 import com.capgemini.library_project.repositories.BookRepository;
 import com.capgemini.library_project.repositories.BorrowRecordRepository;
 import com.capgemini.library_project.repositories.UserRepository;
@@ -58,13 +61,13 @@ public class BorrowRecordServicesImpl implements BorrowRecordServices {
 	    User user = userRepository.findById(userId)
 	            .orElseThrow(() -> {
 	                logger.error("User not found with ID {}", userId);
-	                return new RuntimeException("User not found with ID: " + userId);
+	                return new UserNotFoundException("User with id : " + userId + " not found.");
 	            });
 
 	    Book book = bookRepository.findById(bookId)
 	            .orElseThrow(() -> {
 	                logger.error("Book not found with ID {}", bookId);
-	                return new RuntimeException("Book not found with ID: " + bookId);
+	                return new BookNotFoundException("Book with id " + bookId + " not found");
 	            });
 
 	    if (book.getAvailableCopies() <= 0) {
@@ -105,7 +108,7 @@ public class BorrowRecordServicesImpl implements BorrowRecordServices {
 		return borrowRecordRepository.findById(borrowId)
 				.orElseThrow(() -> {
 					logger.error("No borrow record available for ID {}", borrowId);
-					return new RuntimeException("No Borrow Record available for " + borrowId);
+					return new BorrowRecordNotFoundException("Borrow Record with Id "+borrowId+"not found ");
 				});
 	}
 
@@ -193,7 +196,8 @@ public class BorrowRecordServicesImpl implements BorrowRecordServices {
 		BorrowRecord brecord = borrowRecordRepository.findById(borrowId)
 				.orElseThrow(() -> {
 					logger.error("Borrow record not found for fine calculation");
-					return new RuntimeException("Record not found");
+					return new BorrowRecordNotFoundException("Borrow Record with Id "+borrowId+"not found ");
+					
 				});
 
 		LocalDate dueDate = brecord.getBorrowDate().plusDays(allowedReturnDays);
@@ -232,6 +236,11 @@ public class BorrowRecordServicesImpl implements BorrowRecordServices {
 		logger.info("Updating borrow record ID {}", borrowId);
 		BorrowRecord borrowRecord = getBorrowRecordById(borrowId);
 		
+		if(borrowRecord==null) {
+			throw new BorrowRecordNotFoundException("Borrow Record with Id "+borrowId+"not found ");
+		}
+		
+		
 		String status = updatedBorrowRecord.getBorrowStatus();
         if (status != null && !status.equalsIgnoreCase("Borrowed") &&
             !status.equalsIgnoreCase("Returned") &&
@@ -258,6 +267,10 @@ public class BorrowRecordServicesImpl implements BorrowRecordServices {
 	public void deleteBorrowRecord(Long borrowId) {
 		logger.info("Deleting borrow record with ID {}", borrowId);
 		BorrowRecord record = getBorrowRecordById(borrowId);
+		
+		if(record==null) {
+			throw new BorrowRecordNotFoundException("Borrow Record with Id "+borrowId+"not found ");
+		}
 		borrowRecordRepository.deleteById(borrowId);
 	}
 }
